frontend: Refactor headerAction stuff into headerActionSlice
This moves most of the types and related stuff in close to where
it is used. As well it uses the Redux Toolkit slice way of doing
things which is simplified.
